% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim_voomlike.R
\name{data_sim_voomlike}
\alias{data_sim_voomlike}
\title{Data simulation function}
\usage{
data_sim_voomlike(seed = NULL, maxGSsize = 400, minGSsize = 30,
  beta = 0, do_gs = TRUE, longitudinal = TRUE, mixed_hypothesis = FALSE,
  n = 18, nindiv = 6, ntime = 3)
}
\description{
Adapted from the supplementary material from Law \emph{et al}.
}
\examples{
\dontrun{
set.seed(123)
data_sims <- data_sim_voomlike(maxGSsize = 300)
data_sims <- data_sim_voomlike(maxGSsize = 300, beta = 1.8)
}
}
\references{
Law, C. W., Chen, Y., Shi, W., & Smyth, G. K. (2014). voom: Precision
weights unlock linear model analysis tools for RNA-seq read counts. \emph{Genome
Biology}, 15(2), R29.
}
\keyword{internal}
