% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS_FDR.R
\name{DS_FDR}
\alias{DS_FDR}
\title{Estimating the False Discovery Rate}
\usage{
DS_FDR(gene_scores_perm, gene_scores_obs, n_perm)
}
\arguments{
\item{gene_scores_perm}{a numeric matrix of size \code{G x n} containing the permuted gene-wise scores
for \code{G} genes with \code{n_perm} permutations.}

\item{gene_scores_obs}{a vector of length \code{n} containing the observed gene-wise scores.}

\item{n_perm}{the number of perturbations. Default is \code{1000}.}
}
\value{
A vector of estimating discrete false discovery rates
}
\description{
This function uses the permutation plug-in method to estimate the FDR.
}
\references{
Jun Li and Robert Tibshirani (2011). Finding consistent patterns: A nonparametric approach
for identifying differential expression in RNA-seq data, \emph{Statistical Methods in Medical Research},
0(0) 1–18.
}
