% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dearseq_internal.R
\name{\%^\%}
\alias{\%^\%}
\title{Power for covaroances matrices}
\usage{
x \%^\% n
}
\arguments{
\item{x}{a positive definite symmetric matrix}

\item{n}{a real number}
}
\value{
a matrix of the same dimensions as \code{x}
}
\description{
Compute the power of a positive definite symmetric
}
\keyword{internal}
