% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/London.R
\docType{data}
\name{London}
\alias{London}
\title{A RNA-seq dataset preprocessed using edgeR}
\format{A dataset: \code{London} containing 14,150 genes and 54 samples}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE107991}
}
\usage{
data(London)
}
\description{
A RNA-seq dataset preprocessed using edgeR
}
\examples{
data('London')

}
\references{
Singhania, Akul, \emph{et al.} A modular transcriptional signature
identifies phenotypic heterogeneity of human tuberculosis infection,
\emph{Nature communications} 9.1 (2018): 2308.
}
\keyword{datasets}
