% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/London_raw.R
\docType{data}
\name{London_raw}
\alias{London_raw}
\title{A RNA-seq dataset of raw counts}
\format{A dataset: \code{London_raw} containing 58,051 genes and 54 samples}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE107991}
}
\usage{
data(London_raw)
}
\description{
A RNA-seq dataset of raw counts
}
\examples{
data('London_raw')

}
\references{
Singhania, Akul, \emph{et al.} A modular transcriptional signature
identifies phenotypic heterogeneity of human tuberculosis infection,
\emph{Nature communications} 9.1 (2018): 2308.
}
\keyword{datasets}
