% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nb_sim.R
\name{nb_sim_fn}
\alias{nb_sim_fn}
\title{Computing negative binomial simulations results}
\usage{
nb_sim_fn(n = 250, n_t = 5, beta = 2, nGenes = 1000, type)
}
\description{
Computing negative binomial simulations results
}
\examples{

\dontrun{

all_sim_pvals <- replicate(1000, nb_sim_fn(type = 'standard_null'))
}

}
\keyword{internal}

