% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nonlin_gene_sim.R,
%   R/compute_nonlin_sim.R
\name{nonlin_sim_fn}
\alias{nonlin_sim_fn}
\alias{nonlin_sim_fn}
\title{Computing nonlinear simulations results}
\usage{
nonlin_sim_fn(n = 250, n_t = 5, beta = 2, nGenes = 1000, rho = 0.5,
  re_sd = 1, type)

nonlin_sim_fn(n = 250, n_t = 5, beta = 2, nGenes = 1000, rho = 0.5,
  re_sd = 1, type)
}
\description{
Computing nonlinear simulations results

Computing nonlinear simulations results
}
\examples{

\dontrun{

for (n in c(50,100,150)) {
 betas <- seq(-2, 2, length = 11)
 standard.l <- list()
 for (i in 1:11) {
   standard.sim <- nonlin_sim_fn(type = 'standard',
                                 nGenes = 100,
                                 n = n,
                                 beta = betas[i],
                                 n_t = 5,
                                 re_sd = 1)
   standard.l[[as.character(betas[i])]] <- standard.sim
 }
}

}

\dontrun{
for (n in c(50,100,150)) {
 betas <- seq(-2, 2, length = 11)
 standard.l <- list()
 for (i in 1:11) {
   standard.sim <- nonlin_sim_fn(type = 'standard',
                                 nGenes = 100,
                                 n = n,
                                 beta = betas[i],
                                 n_t = 5,
                                 re_sd = 1)
   standard.l[[as.character(betas[i])]] <- standard.sim
 }
}
}

}
\keyword{internal}
