% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_pvals.R
\name{plot_hist_pvals}
\alias{plot_hist_pvals}
\title{Plotting raw p-values histogram}
\usage{
plot_hist_pvals(pvals, binwidth = 0.02)
}
\arguments{
\item{x}{a vector of raw p-values}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
Display the histogram of raw p-values for diagnostic plots
}
\author{
Boris Hejblum
}
