% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ord_pvals.R
\name{plot_ord_pvals}
\alias{plot_ord_pvals}
\title{Plot of gene-wise p-values}
\usage{
plot_ord_pvals(pvals, signif_threshold = 0.05)
}
\arguments{
\item{pvals}{a vector of length \code{n} containing the raw p-values for
each gene}

\item{signif_threshold}{a value between \code{0} and {1} specifying the 
nominal significance threshold. Default is \code{0.05}.}
}
\value{
a plot of sorted gene-wise p-values
}
\description{
This function prints the sorted exact p-values along with the Benjamini-Hochberg
limit and the 5% threshold
}
