% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_score_h.R
\name{vc_score_h}
\alias{vc_score_h}
\title{Computes variance component test statistic for homogeneous trajectory}
\usage{
vc_score_h(y, x, indiv, phi, w, Sigma_xi = diag(ncol(phi)))
}
\arguments{
\item{y}{a numeric matrix of dim \code{g x n} containing the raw RNAseq counts for g
genes from \code{n} samples}

\item{x}{a numeric design matrix of dim \code{n x p} containing the \code{p} covariates
to be adjusted for}

\item{indiv}{a vector of length \code{n} containing the information for
attributing each sample to one of the studied individuals. Coerced
to be a \code{factor}}

\item{phi}{a numeric design matrix of size \code{n x K} containing the \code{K} variables
to be tested}

\item{w}{a vector of length \code{n} containing the weights for the \code{n}
samples.}

\item{Sigma_xi}{a matrix of size \code{K x K} containing the covariance matrix
of the \code{K} random effects on \code{phi}}
}
\value{
A list with the following elements:\itemize{
  \item \code{score}: approximation of the observed score
  \item \code{q}: TODO
  \item \code{q_ext}: TODO
}
}
\description{
This function computes an approximation of the Variance Component test for a
mixture of \eqn{\chi^{2}}s using Davies method from \code{\link[CompQuadForm]{davies}}
}
\examples{

}
\seealso{
\code{\link[CompQuadForm]{davies}}
}
\keyword{internal}

