% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsFDR.R
\name{dsFDR}
\alias{dsFDR}
\title{Estimating the False Discovery Rate}
\usage{
dsFDR(gene_scores_perm, gene_scores_obs, n_perm)
}
\arguments{
\item{gene_scores_perm}{a numeric matrix of size \code{G x n} containing the permuted gene-wise scores
for \code{G} genes with \code{n_perm} permutations.}

\item{gene_scores_obs}{a vector of length \code{n} containing the observed gene-wise scores.}

\item{n_perm}{the number of perturbations. Default is \code{1000}.}
}
\value{
A vector of estimating discrete false discovery rates
}
\description{
This function uses the permutation plug-in method to estimate the FDR.
}
\references{
J. Li and R. Tibshirani (2013). Finding consistent patterns: A nonparametric approach
for identifying differential expression in RNA-seq data, \emph{Statistical Methods in Medical Research},
22(5): 519-536

J.D. Storey & R. Tibshirani (2003) Statistical significance for genomewide studies.
\emph{Proceedings of the National Academy of Sciences} 100(&6): 9440–9445.
}
