% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgsaseq_internal.R
\name{\%^\%}
\alias{\%^\%}
\title{Power for matrix elements}
\usage{
x \%^\% n
}
\arguments{
\item{x}{a matrix}

\item{n}{a real number}
}
\description{
Set each element of matrix \code{x} to the power \code{n}
}
\keyword{internal}

