% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dearseq.R
\name{plot.dearseq}
\alias{plot.dearseq}
\title{Plot method for dearseq objects}
\usage{
\method{plot}{dearseq}(x, signif_threshold = 0.05, ...)
}
\arguments{
\item{x}{an object of class \code{dear_seq}}

\item{signif_threshold}{a value between \code{0} and {1} specifying the 
nominal significance threshold. Default is \code{0.05}.}

\item{...}{further arguments}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot method for dearseq objects
}
\author{
Boris Hejblum
}
