% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgsaseq-package.R
\docType{package}
\name{tcgsaseq}
\alias{tcgsaseq}
\alias{tcgsaseq-package}
\title{TcGSAseq: a package to perform Time-course Gene Set Analysis of RNA-seq data}
\description{
Gene set analysis of longitudinal RNA-seq data with varaince component score
test acocunting for data heteroscedasticity through precision weights.
}
\details{
\tabular{ll}{
Package: \tab tcgsaseq\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2016-04-30\cr
License:\tab \href{http://www.gnu.org/licenses/gpl-2.0.txt}{GPL-2}\cr
}
The main functions of the TcGSAseq package is \code{\link{tcgsa_seq}}
}
\author{
Boris P. Hejblum, Denis Agniel
--- Maintainer: Boris P. Hejblum
}
\references{
Agniel D, Hejblum BP, Variance component score test for
time-course gene set analysis of longitudinal RNA-seq data, \emph{submitted}, 2016.
}

