% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcGSAseq-package.R
\docType{package}
\name{tcgsaseq}
\alias{tcgsaseq}
\alias{tcgsaseq-package}
\title{tcgsaseq: a package to perform Time-course Gene Set Analysis and General Gene-Wise Analysis of RNA-seq data}
\description{
Analysis of RNA-seq data with variance component score
test accounting for data heteroscedasticity through precision weights.
}
\details{
\tabular{ll}{
Package: \tab tcgsaseq\cr
Type: \tab Package\cr
Version: \tab 1.6.6\cr
Date: \tab 2018-01-15\cr
License:\tab \href{http://www.gnu.org/licenses/gpl-2.0.txt}{GPL-2}\cr
}
The two main functions of the \code{tcgsaseq} package are \code{\link{varseq}} and \code{\link{tcgsa_seq}}.
}
\references{
Agniel D & Hejblum BP (2017). Variance component score test for
time-course gene set analysis of longitudinal RNA-seq data, \emph{Biostatistics},
18(4):589-604. \href{https://doi.org/10.1093/biostatistics/kxx005}{10.1093/biostatistics/kxx005}.
\href{https://arxiv.org/abs/1605.02351}{arXiv:1605.02351}.
}
\author{
Boris P. Hejblum, Denis Agniel
--- Maintainer: Boris P. Hejblum
}
